Parse a Greek excerpt

This function parses a Greek excerpt from the Perseus Digital Library. Parsing includes part of speech, gender, case, mood, voice, tense, person, number, and degree.

parse_excerpt(urn, excerpt)

Arguments

urn
a valid urn from the perseus_catalog object.

excerpt
a valid excerpt, e.g. 5.1-5.5

Value

a data frame

Examples

parse_excerpt("urn:cts:greekLit:tlg0031.tlg002.perseus-grc2", "5.1-5.4")

#> # A tibble: 73 x 12
#>       word     form verse part_of_speech person   number  tense       mood
#>      <chr>    <chr> <chr>          <chr>  <chr>    <chr>  <chr>      <chr>
#>  1     καί      Καὶ   5.1    conjunction   <NA>     <NA>   <NA>       <NA>
#>  2 ἔρχομαι    ἦλθον   5.1           verb  third   plural aorist indicative
#>  3     εἰς      εἰς   5.1    preposition   <NA>     <NA>   <NA>       <NA>
#>  4       ὁ       τὸ   5.1        article   <NA> singular   <NA>       <NA>
#>  5   πέραν    πέραν   5.1         adverb   <NA>     <NA>   <NA>       <NA>
#>  6       ὁ      τῆς   5.1        article   <NA> singular   <NA>       <NA>
#>  7 θάλασσα θαλάσσης   5.1           noun   <NA> singular   <NA>       <NA>
#>  8     εἰς      εἰς   5.1    preposition   <NA>     <NA>   <NA>       <NA>
#>  9       ὁ      τὴν   5.1        article   <NA> singular   <NA>       <NA>
#> 10    χώρα    χώραν   5.1           noun   <NA> singular   <NA>       <NA>
#> # ... with 63 more rows, and 4 more variables: voice <chr>, gender <chr>,
#> #   case <chr>, degree <chr>